Changes to slideshow, bookmarks, etc.
If in slideshow mode, going to the next post will now un-pause the
slideshow if it was paused.
Bookmarking a post will now ensure that the post is not already
bookmarked to avoid duplicates.
Additional error handling if switching to/from using Chrome sync.
Multiple bookmarked posts can be selected and moved/deleted.